Alerts when not login

+++ blogproject/blog/views.py
@@ -3,6 +3,8 @@ from .models import Blog
 from django.utils import timezone
 from django.contrib.auth.models import User
 from django.contrib import auth
+from account.forms import RegisterForm
+from django.contrib import messages
 
 # Create your views here.
 def home(request):
@@ -24,18 +26,23 @@ def new(request):
     return render(request,'new.html')
 
 def create(request):
-    blog=Blog()
-    blog.title=request.POST.get('title',False)
-    blog.body=request.POST.get('body',False)
-    blog.pub_date=timezone.datetime.now()
-    if 'image' in request.FILES:
-        blog.image=request.FILES['image']
-    blog.save()
-    return redirect('/blog/'+str(blog.id))
+    if request.user.is_authenticated:
+        blog=Blog()
+        blog.title=request.POST.get('title',False)
+        blog.body=request.POST.get('body',False)
+        blog.pub_date=timezone.datetime.now()
+        if 'image' in request.FILES:
+            blog.image=request.FILES['image']
+        blog.writer=request.user
+        blog.save()
+        return redirect('/blog/'+str(blog.id))
+    else:
+        return redirect('bloglist')
 
 def delete(request, id):
     delete_blog = Blog.objects.get(id=id)
-    delete_blog.delete()
+    if delete_blog.writer == request.user:
+        delete_blog.delete()
     return redirect('bloglist')
 
 def edit(request, id):
@@ -47,34 +54,8 @@ def update(request, id):
     update_blog.title=request.POST.get('title',False)
     update_blog.body=request.POST.get('body',False)
     update_blog.pub_date=timezone.datetime.now()
-    update_blog.image=request.POST.get('image','')
+    if 'image' in request.FILES:
+        update_blog.image=request.FILES['image']
+    update_blog.writer=request.user
     update_blog.save()
     return redirect('detail',update_blog.id)
-
-'''
-def signup(request):
-    if request.method=='POST':
-        if request.POST['password1']==request.POST['password2']:
-            user=User.objects.create_user(username=request.POST['username'], password=request.POST['password1'])
-            auth.login(request.user)
-            return redirect('home')
-    return render(request,'signup.html')
-
-def login(request):
-    if request.method=='POST':
-        username=request.POST['username']
-        password=request.POST['password']
-        user=auth.authenticate(request,username=username,password=password)
-        if user is not None:
-            auth.login(request,user)
-            return redirect('home')
-        else:
-            return render(request,'login.html',{'error':'username or password is incorrect.'})
-    return render(request,'login.html')
-
-def logout(request):
-    if request.method=='POST':
-        auth.logout(request)
-        redirect('home')
-    return render(request,'login.html')
-'''
